requirements.txt
streamlit 
python-dotenv
pypdf
google-generativeai
PyMuPDF
streamlit-lottie
